added script for music evolution
music will not evolve beyond desolate; unable to discern why. script is attached to Wwiseglobal and is running
