+++ Tera Testing/Assets/Script/Audio/SetAmbienceEvolution.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3501bd2e70bf54a2aadb709c98518906
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
