+++ Tera Testing/Assets/Script/Audio/SetMusicEvolution.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 06197fb72cc85884dadf59d981f2074f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
