+++ Tera Testing/Assets/Wwise/Resources/State.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d79f814ec728a964489b57373d6fb75e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
