+++ Tera Testing/Assets/Wwise/Resources/State/91A293D7-312D-4505-A852-0C4843F48E8F.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0c7ccae51ea2fa34a89195893f9791ea
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
