+++ Tera Testing/Assets/Wwise/Resources/State/AC1B118B-55A5-4C40-8433-E207CDD1315A.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7d95c43301900a84aa3715e7753a774b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
