+++ Tera Testing/Assets/Wwise/Resources/StateGroup.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: efda51260d056b041b43a75ba89548db
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
