+++ Tera Testing/Assets/Wwise/Resources/StateGroup/EB9F3CBE-916B-4602-839C-95AA48193338.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1904f70e5567c480586e3d7843eb0350, type: 3}
+  m_Name: EB9F3CBE-916B-4602-839C-95AA48193338
+  m_EditorClassIdentifier: 
+  objectName: TerraformProgress
+  id: 2315045542
+  guid: EB9F3CBE-916B-4602-839C-95AA48193338
