+++ Tera Testing/Assets/Wwise/Resources/StateGroup/EB9F3CBE-916B-4602-839C-95AA48193338.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f14e7ce5d1b393740845f0b7c441b2e9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
