+++ Tera Testing/Assets/victoria_assets/new planet stuff/newPlanetGrass_MAT.mat
@@ -74,5 +74,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 0
     m_Colors:
-    - _Color: {r: 1, g: 1, b: 1, a: 0}
+    - _Color: {r: 1, g: 1, b: 1, a: 0.18884}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
