+++ Tera Testing/Tera Testing_WwiseProject/.cache/Windows/SFX/SBStat/{18C419C3-7273-4D35-BC56-B32D29CF2496}.stat
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <BankMemory>
-	<Timestamp>1573873127</Timestamp>
+	<Timestamp>1573886072</Timestamp>
 	<DataSize>69702660</DataSize>
-	<FileSize>69707270</FileSize>
+	<FileSize>69707480</FileSize>
 	<DecodedSize>69702452</DecodedSize>
 	<SFXPreFetchSize>0</SFXPreFetchSize>
 	<SFXInMemorySize>41416256</SFXInMemorySize>
