+++ Tera Testing/Tera Testing_WwiseProject/Tera Testing_WwiseProject.Cole Krueger.validationcache
@@ -11,7 +11,7 @@
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Effects\Factory Effects.wwu" DateLow="1573216963" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Events\Default Work Unit.wwu" DateLow="1573873051" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Game Parameters\Default Work Unit.wwu" DateLow="1573216963" DateHigh="0"/>
-	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Interactive Music Hierarchy\Default Work Unit.wwu" DateLow="1573870055" DateHigh="0"/>
+	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Interactive Music Hierarchy\Default Work Unit.wwu" DateLow="1573885891" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Master-Mixer Hierarchy\Default Work Unit.wwu" DateLow="1573216965" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Mixing Sessions\Default Work Unit.wwu" DateLow="1573216965" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Modulators\Default Work Unit.wwu" DateLow="1573216965" DateHigh="0"/>
@@ -20,7 +20,7 @@
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Queries\Factory Queries.wwu" DateLow="1573216968" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\SoundBanks\Default Work Unit.wwu" DateLow="1573873051" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Soundcaster Sessions\Default Work Unit.wwu" DateLow="1573216968" DateHigh="0"/>
-	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\States\Default Work Unit.wwu" DateLow="1573216968" DateHigh="0"/>
+	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\States\Default Work Unit.wwu" DateLow="1573885891" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Switches\Default Work Unit.wwu" DateLow="1573216968" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Tera Testing_WwiseProject.Cole Krueger.wsettings" DateLow="1573872399" DateHigh="0"/>
 	<ValidationCacheItem File="C:\Users\Cole Krueger\Documents\GitHub\TERA_Project\Tera Testing\Tera Testing_WwiseProject\Tera Testing_WwiseProject.wproj" DateLow="1573873051" DateHigh="0"/>
